+++ tubular/misc.py
@@ -2,15 +2,17 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from typing import Any
 
 import narwhals as nw
 import pandas as pd
 from typing_extensions import deprecated
 
-if TYPE_CHECKING:
-    from narwhals.typing import FrameT
-
+from tubular._utils import (
+    _convert_dataframe_to_narwhals,
+    _return_narwhals_or_native_dataframe,
+    block_from_json,
+)
 from tubular.base import BaseTransformer
 
 
@@ -48,7 +50,7 @@ class SetValueTransformer(BaseTransformer):
 
     FITS = False
 
-    jsonable = False
+    jsonable = True
 
     def __init__(
         self,
@@ -74,8 +76,31 @@ class SetValueTransformer(BaseTransformer):
 
         super().__init__(columns=columns, **kwargs)
 
-    @nw.narwhalify
-    def transform(self, X: FrameT) -> FrameT:
+    @block_from_json
+    def to_json(self) -> dict[str, dict[str, Any]]:
+        """Dump transformer to json dict.
+
+        Returns
+        -------
+        dict[str, dict[str, Any]]:
+            jsonified transformer. Nested dict containing levels for attributes
+            set at init and fit.
+
+        Examples
+        --------
+        >>> transformer = SetValueTransformer(columns='a', value=1)
+        >>> transformer.to_json()
+        {'tubular_version': 'dev', 'classname': 'SetValueTransformer', 'init': {'columns': ['a'], 'copy': False, 'verbose': False, 'return_native': True, 'value': 1}, 'fit': {}}
+
+
+        """
+        json_dict = super().to_json()
+
+        json_dict["init"]["value"] = self.value
+
+        return json_dict
+
+    def transform(self, X: pd.DataFrame) -> pd.DataFrame:
         """Set columns to value.
 
         Parameters
@@ -112,10 +137,14 @@ class SetValueTransformer(BaseTransformer):
         └─────┴─────┘
 
         """
-        X = nw.from_native(super().transform(X))
+        X = super().transform(X, return_native_override=False)
 
-        set_value_expression = [nw.lit(self.value).alias(col) for col in self.columns]
-        return X.with_columns(set_value_expression)
+        X = _convert_dataframe_to_narwhals(X)
+
+        for c in self.columns:
+            X = X.with_columns(nw.lit(self.value).alias(c))
+
+        return _return_narwhals_or_native_dataframe(X, self.return_native)
 
 
 # DEPRECATED TRANSFORMERS