+++ CHANGELOG.rst
@@ -33,6 +33,7 @@ Each individual change should have a link to the pull request after the descript
 Changed
 ^^^^^^^
 
+- feat: added `to_json` method for `SetValueTransformer` `#542 <https://github.com/azukds/tubular/issues/542>`_
 - removed SeparatorColumnMixin `#562 <https://github.com/azukds/tubular/issues/562>_`
 
 2.1.0 (30/10/25)